requirement.txt
langchain
ipykernel
python-dotenv
langchain_community
langchain_text_splitters
sentence_transformers
langchain_huggingface
langchain_core
faiss-cpu
langchain_groq
pypdf
streamlit
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb
arxiv
wikipedia